+++ planificador/config.h
@@ -19,8 +19,8 @@ typedef struct {
 	int port;
 	int coordinator_port;
 	t_scheduling_algorithm scheduling_algorithm;
-	double alpha;
-	double initial_estimation;
+	float alpha;
+	float initial_estimation;
 	char* coordinator_ip;
 	char** blocked_keys;
 } t_planificador_config;
